# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backlogg API",
    "description": "Backlogg is a unified catalog API for **movies, series, books and games**,\nseeded from public external APIs (TMDB, Open Library, IGDB) via a nightly sync,\nplus a social layer on top of that catalog.\n\n**Catalog & discovery:** list and detail endpoints per content type, cross-type\n`/search` (with on-demand external fallback), `/genres`, `/trending`, `/people`\nand similar-item recommendations.\n\n**Social layer:** account management with refresh tokens and account recovery\n(`/auth`, `/users`), ratings & reviews, per-user library/backlog, curated\n`/lists`, `/follows`, an activity `/feed`, `/notifications` and personalized\n`/recommendations`.\n\n**Platform:** rate limiting, structured observability, Prometheus `/metrics`\nand response caching.\n\n## Authentication\n\n- **User endpoints** use `Authorization: Bearer <access_token>`. Obtain the\n  token pair from `POST /auth/login`; renew it with `POST /auth/refresh`.\n- **Admin endpoints** (`/admin/*`) require the `X-API-Key` header.\n\nEverything else is public. Direct messaging between users is out of scope.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/series": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "List series",
        "description": "Paginated list of catalogued series (no external fallback); genre filter + sort.",
        "operationId": "list_series_v1_series_get",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by genre slug",
              "title": "Genre"
            },
            "description": "Filter by genre slug"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SeriesSortEnum",
              "description": "Sort order",
              "default": "rating_desc"
            },
            "description": "Sort order"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/series/{slug}/similar": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "Similar series",
        "description": "Up to 10 similar series (TMDB). New items are persisted locally.",
        "operationId": "get_similar_series_v1_series__slug__similar_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarSeriesListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/series/{slug}/ratings": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "List series ratings",
        "description": "Public, paginated ratings & reviews for a series, newest first.",
        "operationId": "list_series_ratings_v1_series__slug__ratings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/series/{slug}/rating": {
      "put": {
        "tags": [
          "series"
        ],
        "summary": "Rate a series",
        "description": "Upsert the caller's score/review (full replace); recomputes aggregates. Auth.",
        "operationId": "rate_series_v1_series__slug__rating_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "series"
        ],
        "summary": "Delete own series rating",
        "description": "Remove the caller's rating and recompute aggregates. Requires auth.",
        "operationId": "delete_series_rating_v1_series__slug__rating_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/series/{slug}/library": {
      "put": {
        "tags": [
          "series"
        ],
        "summary": "Set series library status",
        "description": "Upsert the caller's backlog status for this series. Requires auth.",
        "operationId": "set_series_library_v1_series__slug__library_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryEntryIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "series"
        ],
        "summary": "Remove series from library",
        "description": "Delete the caller's library entry for this series. Requires auth.",
        "operationId": "delete_series_library_v1_series__slug__library_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/series/{slug}": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "Get series detail",
        "description": "Full series detail; on-demand fallback. Auth optional (adds viewer_status).",
        "operationId": "get_series_v1_series__slug__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibraryEntryIn": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LibraryStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "LibraryEntryIn",
        "description": "Request body for PUT /{type}/{slug}/library."
      },
...
      "LibraryStatus": {
        "type": "string",
        "enum": [
          "want",
          "in_progress",
          "completed",
          "dropped"
        ],
        "title": "LibraryStatus",
        "description": "Backlog status values accepted in request bodies and filters."
      },
      "LibraryStatusOut": {
        "properties": {
          "item_type": {
            "type": "string",
            "title": "Item Type"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "item_type",
          "slug",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "LibraryStatusOut",
        "description": "Response of PUT /{type}/{slug}/library \u2014 the caller's own entry.",
        "example": {
          "created_at": "2026-05-25T18:04:11Z",
          "item_type": "MOVIE",
          "slug": "dune-2021",
          "status": "want",
          "updated_at": "2026-05-25T18:04:11Z"
        }
      },
...
      "RatingAuthorOut": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "username",
          "display_name",
          "avatar_url"
        ],
        "title": "RatingAuthorOut"
      },
      "RatingIn": {
        "properties": {
          "score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "review_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Text"
          }
        },
        "type": "object",
        "title": "RatingIn",
        "description": "Request body for PUT /{type}/{slug}/rating \u2014 full replace (upsert)."
      },
      "RatingListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RatingOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "title": "RatingListOut"
      },
      "RatingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user": {
            "$ref": "#/components/schemas/RatingAuthorOut"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "review_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Text"
          },
          "like_count": {
            "type": "integer",
            "title": "Like Count"
          },
          "liked_by_viewer": {
            "type": "boolean",
            "title": "Liked By Viewer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user",
          "score",
          "review_text",
          "like_count",
          "liked_by_viewer",
          "created_at",
          "updated_at"
        ],
        "title": "RatingOut"
      },
...
      "SeriesGenreOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "SeriesGenreOut"
      },
      "SeriesListItemOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "rating_external": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating External"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "poster_url",
          "release_date",
          "rating_external",
          "genres"
        ],
        "title": "SeriesListItemOut"
      },
      "SeriesListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SeriesListItemOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ],
        "title": "SeriesListOut"
      },
      "SeriesOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "original_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "overview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overview"
          },
          "first_air_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Air Date"
          },
          "last_air_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Air Date"
          },
          "number_of_seasons": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number Of Seasons"
          },
          "number_of_episodes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number Of Episodes"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "original_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Language"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "backdrop_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backdrop Url"
          },
          "rating_external": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating External"
          },
          "rating_count_external": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Count External"
          },
          "rating_internal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Internal"
          },
          "rating_count_internal": {
            "type": "integer",
            "title": "Rating Count Internal"
          },
          "genres": {
            "items": {
              "$ref": "#/components/schemas/SeriesGenreOut"
            },
            "type": "array",
            "title": "Genres"
          },
          "credits": {
            "items": {
              "$ref": "#/components/schemas/backlogg__shared__schemas__CreditOut"
            },
            "type": "array",
            "title": "Credits",
            "default": []
          },
          "viewer_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewer Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "original_title",
          "slug",
          "overview",
          "first_air_date",
          "last_air_date",
          "number_of_seasons",
          "number_of_episodes",
          "status",
          "original_language",
          "poster_url",
          "backdrop_url",
          "rating_external",
          "rating_count_external",
          "rating_internal",
          "rating_count_internal",
          "genres"
        ],
        "title": "SeriesOut"
      },
      "SeriesSortEnum": {
        "type": "string",
        "enum": [
          "rating_desc",
          "rating_asc",
          "date_desc",
          "date_asc",
          "title_asc"
        ],
        "title": "SeriesSortEnum"
      },
...
      "SimilarSeriesListOut": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SimilarSeriesOut"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "SimilarSeriesListOut"
      },
      "SimilarSeriesOut": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "rating_external": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating External"
          }
        },
        "type": "object",
        "required": [
          "title",
          "slug",
          "poster_url",
          "release_date",
          "rating_external"
        ],
        "title": "SimilarSeriesOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "backlogg__shared__schemas__CreditOut": {
        "properties": {
          "person_name": {
            "type": "string",
            "title": "Person Name"
          },
          "person_slug": {
            "type": "string",
            "title": "Person Slug"
          },
          "profile_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Url"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "character_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Character Name"
          },
          "billing_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billing Order"
          }
        },
        "type": "object",
        "required": [
          "person_name",
          "person_slug",
          "profile_url",
          "role",
          "character_name",
          "billing_order"
        ],
        "title": "CreditOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}